Refactor parsing into its own file
